requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
pyaes==1.6.1
PySocks==1.7.1
python-dotenv==1.2.1
pywin32==311
requests==2.32.5
urllib3==2.5.0
